requirements.txt
Django==1.4.3
Fabric==1.4.3
South==0.7.6
django-debug-toolbar==0.9.4
raven==2.0.6
django-devserver==0.4.0
dj-database-url==0.2.1
gunicorn==0.14.6
django-nose==1.1
coverage==3.5.3
nose==1.2.1
django-allauth==0.15.0
django-grappelli==2.4.3
django-taggit==0.9.3
sorl-thumbnail==11.12
django-taggit-templatetags==0.4.6dev
django-filebrowser==3.5.1
django-rosetta==0.7.1
django-bootstrap-form==2.0.5
